+++ tests/Unit/Employees/EmployeeUnitTest.php
@@ -29,4 +29,17 @@ class EmployeeUnitTest extends TestCase
             $this->assertEquals($userRole->name, $role->name);
         });
     }
+    /** @test */
+    public function it_can_attach_or_detach_the_employee_role_()
+    {
+        $employee = factory(Employee::class)->create();
+        $role = factory(Role::class)->create();
+
+        $employee->roles()->attach($role);
+        $this->assertTrue($employee->hasRole($role->name));
+
+        $employee->roles()->detach($role);
+        $this->assertFalse($employee->hasRole($role->name));
+
+    }
 }
